+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- CAPA: check for deletion timestamp on the Cluster CR.
+
 ## [0.25.0] - 2024-02-13
 
 ### Fixed